requirements.txt
browser_cookie3==0.20.1
bs4==0.0.2
matplotlib~=3.8.1
numpy~=1.26.3
pandas~=2.2.3
python-dotenv~=1.0.1
requests~=2.32.3
rich~=13.9.4
seaborn~=0.13.1
tqdm~=4.67.1
